* удалены SecurityUser и UserDetailsServiceImpl * класс User implements UserDetails * UserService implements UserDetailsService * autowired UserService в SecurityConfig
